# Reorder vertices in the order in which we encounter them in cell loops.