--- a/tests/bits/find_cells_adjacent_to_vertex_3.output
+++ b/tests/bits/find_cells_adjacent_to_vertex_3.output
@@ -11,40 +11,40 @@ DEAL::   0.7
 DEAL::Vertex 1 at 1.00000 0.00000: 2 cells
 DEAL::   0.0
 DEAL::   0.7
-DEAL::Vertex 2 at 0.923880 0.382683: 1 cells
-DEAL::   0.0
-DEAL::Vertex 3 at 0.707107 0.707107: 2 cells
+DEAL::Vertex 2 at 0.707107 0.707107: 2 cells
 DEAL::   0.0
 DEAL::   0.1
-DEAL::Vertex 4 at 0.382683 0.923880: 1 cells
-DEAL::   0.1
-DEAL::Vertex 5 at 6.12323e-17 1.00000: 2 cells
+DEAL::Vertex 3 at 0.923880 0.382683: 1 cells
+DEAL::   0.0
+DEAL::Vertex 4 at 6.12323e-17 1.00000: 2 cells
 DEAL::   0.1
 DEAL::   0.2
-DEAL::Vertex 6 at -0.382683 0.923880: 1 cells
-DEAL::   0.2
-DEAL::Vertex 7 at -0.707107 0.707107: 2 cells
+DEAL::Vertex 5 at 0.382683 0.923880: 1 cells
+DEAL::   0.1
+DEAL::Vertex 6 at -0.707107 0.707107: 2 cells
 DEAL::   0.2
 DEAL::   0.3
-DEAL::Vertex 8 at -0.923880 0.382683: 1 cells
-DEAL::   0.3
-DEAL::Vertex 9 at -1.00000 1.22465e-16: 2 cells
+DEAL::Vertex 7 at -0.382683 0.923880: 1 cells
+DEAL::   0.2
+DEAL::Vertex 8 at -1.00000 1.22465e-16: 2 cells
 DEAL::   0.3
 DEAL::   0.4
-DEAL::Vertex 10 at -0.923880 -0.382683: 1 cells
-DEAL::   0.4
-DEAL::Vertex 11 at -0.707107 -0.707107: 2 cells
+DEAL::Vertex 9 at -0.923880 0.382683: 1 cells
+DEAL::   0.3
+DEAL::Vertex 10 at -0.707107 -0.707107: 2 cells
 DEAL::   0.4
 DEAL::   0.5
-DEAL::Vertex 12 at -0.382683 -0.923880: 1 cells
-DEAL::   0.5
-DEAL::Vertex 13 at -1.83697e-16 -1.00000: 2 cells
+DEAL::Vertex 11 at -0.923880 -0.382683: 1 cells
+DEAL::   0.4
+DEAL::Vertex 12 at -1.83697e-16 -1.00000: 2 cells
 DEAL::   0.5
 DEAL::   0.6
-DEAL::Vertex 14 at 0.382683 -0.923880: 1 cells
-DEAL::   0.6
-DEAL::Vertex 15 at 0.707107 -0.707107: 2 cells
+DEAL::Vertex 13 at -0.382683 -0.923880: 1 cells
+DEAL::   0.5
+DEAL::Vertex 14 at 0.707107 -0.707107: 2 cells
 DEAL::   0.6
 DEAL::   0.7
+DEAL::Vertex 15 at 0.382683 -0.923880: 1 cells
+DEAL::   0.6
 DEAL::Vertex 16 at 0.923880 -0.382683: 1 cells
 DEAL::   0.7