--- a/tests/codim_one/extract_boundary_mesh_05.output
+++ b/tests/codim_one/extract_boundary_mesh_05.output
@@ -5,14 +5,14 @@ DEAL::vertex: 0, global: 0 index: 0
 DEAL::vertex: 1, global: 1 index: 1
 DEAL::
 DEAL::Cell: 0.1
-DEAL::vertex: 0, global: 3 index: 2
-DEAL::vertex: 1, global: 2 index: 3
+DEAL::vertex: 0, global: 2 index: 2
+DEAL::vertex: 1, global: 3 index: 3
 DEAL::
 DEAL::Cell: 0.2
-DEAL::vertex: 0, global: 2 index: 3
+DEAL::vertex: 0, global: 3 index: 3
 DEAL::vertex: 1, global: 0 index: 0
 DEAL::
 DEAL::Cell: 0.3
 DEAL::vertex: 0, global: 1 index: 1
-DEAL::vertex: 1, global: 3 index: 2
+DEAL::vertex: 1, global: 2 index: 2
 DEAL::