--- a/tests/dofs/dof_accessor_01.output
+++ b/tests/dofs/dof_accessor_01.output
@@ -1,4 +1,4 @@
 
-DEAL::3 3
+DEAL::1 1
 DEAL::40 40
 DEAL::281 281