--- a/tests/fe/nedelec_non_rect_2d.with_umfpack=on.output
+++ b/tests/fe/nedelec_non_rect_2d.with_umfpack=on.output
@@ -1,4 +1,4 @@
 
-DEAL::p=0 L2_error: 0.864275
-DEAL::p=1 L2_error: 0.0650158
-DEAL::p=2 L2_error: 2.32125e-15
+DEAL::p=0 L2_error: 0.808027
+DEAL::p=1 L2_error: 0.0985218
+DEAL::p=2 L2_error: 3.57575e-15