--- a/tests/grid/grid_in_3d_02.output
+++ b/tests/grid/grid_in_3d_02.output
@@ -1,3 +1,3 @@
 
 DEAL::  51823 active cells
-DEAL::  hash=797857176
+DEAL::  hash=798125619