Here is a draft attempt at reordering vertices in the order in which we encounter them in loops over all cells, in an attempt at improving cache locality. This is perhaps not relevant for a program like step-6, but in ASPECT we probably perform several dozen loops, perhaps hundreds, over all cells and access vertices in most of them. Having them ordered in the order in which we read them is useful in this case.

This patch has the potential to massively upend the test suite in all of the places where the order of vertices matters, or that simply output vertex information in some way. Let's just see what happens here when we run the test suite :-)

Before this gets merged, we should also do some performance studies on real codes. For the moment, this is simply a draft patch.

This fixes #15044.