@tamiko 2023-07-09
A couple of simplex mesh tests are currently failing:
```
The following tests FAILED:
 	 36 - simplex/convert_hypercube_to_simplex_mesh_01.debug (Failed)
 	 37 - simplex/convert_hypercube_to_simplex_mesh_02.debug (Failed)
 	 64 - simplex/grid_in_msh.debug (Failed)
 	 65 - simplex/grid_in_vtk.debug (Failed)
 	 69 - simplex/grid_out_msh.debug (Failed)
 	106 - simplex/refinement_01.debug (Failed)
 	109 - simplex/refinement_04.debug (Failed)
```