@bangerth 2023-07-09
"880+102 tests" -- yes, that's what I was afraid of. At the same time, the fact that 3200 tests actually succeeded means that it can't be *completely* broken :-)