@bangerth 2024-04-08
I'm down to ~15 failures that mostly have to do with some algorithms in `TriaDescription`. I will get to those at some point, but in the meantime I thought I'd try and assess whether this whole approach is actually worthwhile.

I took step-6 as an example, adding a few refinement cycles to assess the outcome. Here is running things in 2d before the patch (3 runs to see what the variability is):
```
Cycle 7:
   Number of active cells:       114866
   Number of degrees of freedom: 516565


+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      51.1s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.388s |      0.76% |
| output results                  |         1 |     0.837s |       1.6% |
| refine_grid                     |         1 |     0.515s |         1% |
| setup                           |         1 |     0.414s |      0.81% |
| solve                           |         1 |        49s |        96% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      51.8s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.383s |      0.74% |
| output results                  |         1 |     0.801s |       1.5% |
| refine_grid                     |         1 |     0.519s |         1% |
| setup                           |         1 |     0.425s |      0.82% |
| solve                           |         1 |      49.7s |        96% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      48.9s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.417s |      0.85% |
| output results                  |         1 |     0.814s |       1.7% |
| refine_grid                     |         1 |     0.513s |         1% |
| setup                           |         1 |     0.437s |      0.89% |
| solve                           |         1 |      46.8s |        96% |
+---------------------------------+-----------+------------+------------+
```
So without the patch the best times are:
* assemble: 0.383
* output results: 0.801
* refine grid: 0.515
* setup: 0.414


And these are the results with the patch:
```
Cycle 7:
   Number of active cells:       114866
   Number of degrees of freedom: 516565


+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      52.2s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.501s |      0.96% |
| output results                  |         1 |     0.806s |       1.5% |
| refine_grid                     |         1 |     0.517s |      0.99% |
| setup                           |         1 |     0.484s |      0.93% |
| solve                           |         1 |      49.9s |        96% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      49.3s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.478s |      0.97% |
| output results                  |         1 |     0.826s |       1.7% |
| refine_grid                     |         1 |     0.523s |       1.1% |
| setup                           |         1 |     0.521s |       1.1% |
| solve                           |         1 |      46.9s |        95% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      53.1s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assemble                        |         1 |     0.475s |      0.89% |
| output results                  |         1 |     0.818s |       1.5% |
| refine_grid                     |         1 |     0.523s |      0.98% |
| setup                           |         1 |      0.55s |         1% |
| solve                           |         1 |      50.7s |        96% |
+---------------------------------+-----------+------------+------------+
```

So with the patch the best times are:
* assemble: 0.475
* output results: 0.806
* refine grid: 0.517
* setup: 0.484

I have to admit that I'm a bit stumped. It does not make much sense to me that *all* of these operations should be slower. Instead, I would have expected that we end up with fewer cache misses, and better speed. I will report back on further experiments later.