@bangerth 2024-04-10
step-6 is perhaps not a great test because the solver is so dominant. I tried this again with step-40, for which I get the following (using the same format).

**Without the patch**

```
Before:
Cycle 7:
   Number of active cells:       1495342
   Number of degrees of freedom: 6617939
   Solved in 16 iterations.


+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      83.7s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |       5.5s |       6.6% |
| output                          |         1 |      4.39s |       5.3% |
| refine                          |         1 |      9.84s |        12% |
| setup                           |         1 |      8.24s |       9.9% |
| solve                           |         1 |      55.7s |        67% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      83.6s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |      5.45s |       6.5% |
| output                          |         1 |      4.44s |       5.3% |
| refine                          |         1 |      9.61s |        11% |
| setup                           |         1 |      8.28s |       9.9% |
| solve                           |         1 |      55.9s |        67% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      83.5s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |       5.5s |       6.6% |
| output                          |         1 |       4.4s |       5.3% |
| refine                          |         1 |      9.73s |        12% |
| setup                           |         1 |      8.22s |       9.8% |
| solve                           |         1 |      55.7s |        67% |
+---------------------------------+-----------+------------+------------+

Best:
assembly 5.45s
output 4.39s
refine 9.61s
setup 8.22s
```

**With the patch**

```
Cycle 7:
   Number of active cells:       1495342
   Number of degrees of freedom: 6617939
   Solved in 16 iterations.


+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      82.8s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |      5.48s |       6.6% |
| output                          |         1 |      4.27s |       5.2% |
| refine                          |         1 |      9.77s |        12% |
| setup                           |         1 |      8.14s |       9.8% |
| solve                           |         1 |      55.1s |        67% |
+---------------------------------+-----------+------------+------------+

+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      82.8s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |      5.51s |       6.7% |
| output                          |         1 |      4.31s |       5.2% |
| refine                          |         1 |      9.77s |        12% |
| setup                           |         1 |      8.19s |       9.9% |
| solve                           |         1 |        55s |        66% |
+---------------------------------+-----------+------------+------------+
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      82.7s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| assembly                        |         1 |      5.53s |       6.7% |
| output                          |         1 |      4.29s |       5.2% |
| refine                          |         1 |      9.72s |        12% |
| setup                           |         1 |      8.16s |       9.9% |
| solve                           |         1 |        55s |        66% |
+---------------------------------+-----------+------------+------------+

Best:
assembly 5.48s  (+0.03s = +1%)
output   4.27s  (-0.12s = -3%)
refine   9.72s  (+0.11s = +1%)
setup    8.14s  (=0.08s = -1%)
```

So some operations are slightly faster, others are slightly slower. (The solver appears to be 1.5% faster with the patch, which I cannot reconcile because I believe that the matrix and right hand side should be exactly the same.)

It will be interesting to see what @gassmoeller finds on his benchmark.